+++ packages/store/src/internal/state-operations.ts
@@ -1,10 +1,11 @@
-import { Injectable } from '@angular/core';
+import { Injectable, isDevMode } from '@angular/core';
 
 import { StateOperations } from '../internal/internals';
 import { InternalDispatcher } from '../internal/dispatcher';
 import { StateStream } from './state-stream';
 import { NgxsConfig } from '../symbols';
 import { deepFreeze } from '../utils/freeze';
+import { isAngularInTestMode } from '../utils/angular';
 
 /**
  * State Context factory class
@@ -12,8 +13,12 @@ import { deepFreeze } from '../utils/freeze';
  */
 @Injectable()
 export class InternalStateOperations {
-  constructor(private _stateStream: StateStream, private _dispatcher: InternalDispatcher, private _config: NgxsConfig) {
-    this.checkDevelopmentMode();
+  constructor(
+    private _stateStream: StateStream,
+    private _dispatcher: InternalDispatcher,
+    private _config: NgxsConfig
+  ) {
+    this.verifyDevMode();
   }
 
   /**
@@ -22,8 +27,8 @@ export class InternalStateOperations {
   getRootStateOperations(): StateOperations<any> {
     const rootStateOperations = {
       getState: () => this._stateStream.getValue(),
-      setState: newState => this._stateStream.next(newState),
-      dispatch: actions => this._dispatcher.dispatch(actions)
+      setState: (newState: any) => this._stateStream.next(newState),
+      dispatch: (actions: any[]) => this._dispatcher.dispatch(actions)
     };
 
     if (this._config.developmentMode) {
@@ -33,12 +38,25 @@ export class InternalStateOperations {
     return rootStateOperations;
   }
 
-  private checkDevelopmentMode() {
-    if (this._config.developmentMode) {
+  private verifyDevMode() {
+    if (isAngularInTestMode()) return;
+
+    const isNgxsDevMode = this._config.developmentMode;
+    const isNgDevMode = isDevMode();
+    const incorrectProduction = !isNgDevMode && isNgxsDevMode;
+    const incorrectDevelopment = isNgDevMode && !isNgxsDevMode;
+    const example = 'NgxsModule.forRoot(states, { developmentMode: !environment.production })';
+
+    if (incorrectProduction) {
       console.warn(
-        'NGXS is running in the development mode.\n',
-        'Set developmentMode to false on the NgxsModule options to enable the production mode.\n',
-        'NgxsModule.forRoot(states, { developmentMode: !environment.production })'
+        'Angular is running in production mode but NGXS is still running in the development mode!\n',
+        'Please set developmentMode to false on the NgxsModule options when in production mode.\n',
+        example
+      );
+    } else if (incorrectDevelopment) {
+      console.warn(
+        'RECOMMENDATION: Set developmentMode to true on the NgxsModule when Angular is running in development mode.\n',
+        example
       );
     }
   }