+++ packages/store/tests/selector.spec.ts
@@ -15,7 +15,7 @@ describe('Selector', () => {
   })
   class MyState {
     @Selector()
-    static foo(state) {
+    static foo(state: any) {
       return state.foo;
     }
   }
@@ -29,19 +29,19 @@ describe('Selector', () => {
   })
   class MyState2 {
     @Selector([MyState.foo])
-    static foo(myState2, myStateFoo) {
+    static foo(myState2: any, myStateFoo: any) {
       return myState2.foo + myStateFoo;
     }
 
     @Selector([MyState2.foo])
-    static fooBar(myState2, foo) {
+    static fooBar(myState2: any, foo: any) {
       return foo + myState2.bar;
     }
   }
 
   class MetaSelector {
     @Selector([MyState.foo])
-    static foo(myState) {
+    static foo(myState: any) {
       return myState;
     }
   }
@@ -100,7 +100,7 @@ describe('Selector', () => {
 
     describe('(memoization)', () => {
       it('should memoize the last result', async(() => {
-        const selectorCalls = [];
+        const selectorCalls: string[] = [];
 
         @State<any>({
           name: 'counter',
@@ -111,13 +111,13 @@ describe('Selector', () => {
         })
         class TestState {
           @Selector()
-          static foo(state) {
+          static foo(state: any) {
             selectorCalls.push('foo');
             return state.foo;
           }
 
           @Selector()
-          static bar(state) {
+          static bar(state: any) {
             selectorCalls.push('bar');
             return state.bar;
           }
@@ -137,7 +137,7 @@ describe('Selector', () => {
       }));
 
       it('should memoize the last result of an inner function', async(() => {
-        const selectorCalls = [];
+        const selectorCalls: string[] = [];
 
         @State<any>({
           name: 'counter',
@@ -148,7 +148,7 @@ describe('Selector', () => {
         })
         class TestState {
           @Selector()
-          static foo(state) {
+          static foo(state: any) {
             selectorCalls.push('foo[outer]');
             return () => {
               selectorCalls.push('foo[inner]');
@@ -180,7 +180,10 @@ describe('Selector', () => {
       });
 
       const store: Store = TestBed.get(Store);
-      const selector = createSelector([MyState], state => state.foo);
+      const selector = createSelector(
+        [MyState],
+        (state: any) => state.foo
+      );
       const slice = store.selectSnapshot(selector);
       expect(slice).toBe('Hello');
     }));
@@ -191,7 +194,10 @@ describe('Selector', () => {
       });
 
       const store: Store = TestBed.get(Store);
-      const selector = createSelector([MyState.foo], state => state);
+      const selector = createSelector(
+        [MyState.foo],
+        (state: any) => state
+      );
       const slice = store.selectSnapshot(selector);
       expect(slice).toBe('Hello');
     }));
@@ -203,7 +209,10 @@ describe('Selector', () => {
 
       const store: Store = TestBed.get(Store);
       const myState = store.selectSnapshot(<any>MyState);
-      const selector = createSelector([MyState], state => state.foo);
+      const selector = createSelector(
+        [MyState],
+        (state: any) => state.foo
+      );
       const slice = selector(myState);
       expect(slice).toBe('Hello');
     }));
@@ -214,14 +223,17 @@ describe('Selector', () => {
       });
 
       const store: Store = TestBed.get(Store);
-      const selector = createSelector([MyState, MyState.foo], (state, foo) => state.foo + foo);
+      const selector: any = createSelector(
+        [MyState, MyState.foo],
+        (state: any, foo: any) => state.foo + foo
+      );
       const slice = store.selectSnapshot(selector);
       expect(slice).toBe('HelloHello');
     }));
 
     describe('(memoization)', () => {
       it('should memoize the last result', async(() => {
-        const selectorCalls = [];
+        const selectorCalls: string[] = [];
 
         @State<any>({
           name: 'counter',
@@ -238,14 +250,20 @@ describe('Selector', () => {
 
         const store: Store = TestBed.get(Store);
 
-        const fooSelector = createSelector([TestState], state => {
-          selectorCalls.push('foo');
-          return state.foo;
-        });
-        const barSelector = createSelector([TestState], state => {
-          selectorCalls.push('bar');
-          return state.bar;
-        });
+        const fooSelector = createSelector(
+          [TestState],
+          (state: any) => {
+            selectorCalls.push('foo');
+            return state.foo;
+          }
+        );
+        const barSelector = createSelector(
+          [TestState],
+          (state: any) => {
+            selectorCalls.push('bar');
+            return state.bar;
+          }
+        );
         store.selectSnapshot(fooSelector);
         store.selectSnapshot(fooSelector);
         store.selectSnapshot(barSelector);
@@ -255,7 +273,7 @@ describe('Selector', () => {
       }));
 
       it('should memoize the last result of an inner function', async(() => {
-        const selectorCalls = [];
+        const selectorCalls: string[] = [];
 
         @State<any>({
           name: 'counter',
@@ -271,13 +289,16 @@ describe('Selector', () => {
         });
 
         const store: Store = TestBed.get(Store);
-        const fooSelector = createSelector([TestState], state => {
-          selectorCalls.push('foo[outer]');
-          return () => {
-            selectorCalls.push('foo[inner]');
-            return state.foo;
-          };
-        });
+        const fooSelector = createSelector(
+          [TestState],
+          (state: any) => {
+            selectorCalls.push('foo[outer]');
+            return () => {
+              selectorCalls.push('foo[inner]');
+              return state.foo;
+            };
+          }
+        );
         store.selectSnapshot(fooSelector);
         store.selectSnapshot(fooSelector)();
         const fn = store.selectSnapshot(fooSelector);